requirements.txt
# Web Framework and Extensions
Flask==2.0.3
Flask-SocketIO==5.3.6
Flask-Cors==5.0.0
Flask-Login==0.6.3
gunicorn==21.2.0
gevent==23.9.1

# Database
psycopg2-binary==2.9.9
psycopg2-pool==1.1

# Message Queue & Background Tasks
redis==4.5.5
celery==5.3.6

# API Clients
openai==1.6.1
twilio==8.10.0
cachetools==5.3.2
google-api-python-client==2.108.0
google-auth==2.23.4

# Natural Language Processing
langdetect==1.0.9
python-dateutil==2.9.0.post0

# Utilities
requests==2.31.0
tenacity==8.2.3
python-dotenv==1.0.0
concurrent-log-handler==0.9.25